requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
tqdm
scikit-learn
python-dotenv